Refactor authentication handling: replace useAuth hook with useAuthQueries for improved API interactions, implement centralized error handling, and enhance user experience with notifications; remove deprecated AuthProvider and useAuth hooks.
