+++ src/hooks/userMenuItems.tsx
@@ -1,11 +1,11 @@
 // menuItems.tsx
 
 import webRoutes from '@/lib/web.routes';
-import { useAuth } from '@/modules/auth/hooks/useAuth';
 import AccountCircleIcon from '@mui/icons-material/AccountCircle';
 import ExitToAppIcon from '@mui/icons-material/ExitToApp';
 import { useNavigate } from 'react-router';
 import { Roles } from '@/constants/roles';
+import useAuthQueries from '@/modules/auth/hooks/useAuthQueries';
 
 interface MenuItemType {
   text: string;
@@ -24,12 +24,12 @@ const createMenuItem = (text: string, icon: React.ReactNode, roles: Roles[], onC
 });
 
 export const useUserMenuItems = (userRole: Roles): MenuItemType[] => {
-  const auth = useAuth();
+  const { logoutMutation } = useAuthQueries();
   const navigate = useNavigate();
 
   const allMenuItems: MenuItemType[] = [
     createMenuItem('Profile', <AccountCircleIcon fontSize="small" />, [Roles.All], () => navigate(webRoutes.backoffice.user.profile), webRoutes.backoffice.user.profile),
-    createMenuItem('Logout', <ExitToAppIcon fontSize="small" />, [Roles.All], () => auth.logout()),
+    createMenuItem('Logout', <ExitToAppIcon fontSize="small" />, [Roles.All], () => logoutMutation.mutateAsync()),
   ];
 
   const filterItemsByRole = <T extends { roles: Roles[] }>(items: T[], role: Roles): T[] => {
