--- src/modules/auth/hooks/useAuth.ts
@@ -1,10 +1,0 @@
-import { useContext } from "react";
-import { AuthContext } from "../context/AuthContext";
-
-export const useAuth = () => {
-  const context = useContext(AuthContext);
-  if (!context) {
-    throw new Error("useAuth debe usarse dentro de un AuthProvider");
-  }
-  return context;
-};
