+++ src/modules/auth/pages/ForgotPasswordPopup.tsx
@@ -1,11 +1,10 @@
-import { z } from "zod";
-import { zodResolver } from "@hookform/resolvers/zod";
-import { useErrorStore, useLoaderStore, useNotificationStore } from "@/stores";
-import { useAuth } from "../hooks/useAuth";
-import { FormProvider, SubmitHandler, useForm } from "react-hook-form";
-import { data } from "react-router";
-import { Box, Button, Dialog, DialogActions, DialogContent, DialogTitle, Typography } from "@mui/material";
-import TextFieldForm from "@/components/ui/atoms/TextFieldForm";
+import { z } from 'zod';
+import { zodResolver } from '@hookform/resolvers/zod';
+import { useNotificationStore } from '@/stores';
+import { FormProvider, SubmitHandler, useForm } from 'react-hook-form';
+import { Box, Button, Dialog, DialogActions, DialogContent, DialogTitle, Typography } from '@mui/material';
+import TextFieldForm from '@/components/ui/atoms/TextFieldForm';
+import useAuthQueries from '../hooks/useAuthQueries';
 
 interface ForgotPasswordProps {
   open: boolean;
@@ -17,43 +16,37 @@ interface IFormInput {
 }
 
 const validationSchema = z.object({
-  email: z.string().email("Invalid email address").nonempty("Email is required"),
+  email: z.string().email('Invalid email address').nonempty('Email is required'),
 });
 
 const ForgotPasswordPopup: React.FC<ForgotPasswordProps> = ({ open, onClose }) => {
   const { notification, setNotification, clearNotification } = useNotificationStore();
-  const { loading, setLoading } = useLoaderStore();
-
-  const { forgotPassword } = useAuth();
+  const { forgotPasswordMutation } = useAuthQueries();
+  const loading = forgotPasswordMutation.isPending;
 
   const methods = useForm({
-    mode: "all",
+    mode: 'all',
     resolver: zodResolver(validationSchema),
   });
 
-  const {
-    handleSubmit,
-    formState: { errors },
-    reset,
-  } = methods;
+  const { handleSubmit } = methods;
 
   const onSubmit: SubmitHandler<IFormInput> = async (data) => {
-    setLoading(true);
     clearNotification();
-    try {
-      await forgotPassword(data.email);
-      setNotification({
-        message: "If a user account exists with the entered email, we will send an email with further instructions to reset the password.",
-        type: "success",
-      });
-    } catch (error) {
-      setNotification({
-        message: "An error occurred. Please try again later",
-        type: "error",
-      });
-    } finally {
-      setLoading(false);
-    }
+    forgotPasswordMutation.mutateAsync(data.email, {
+      onSuccess: () => {
+        setNotification({
+          message: 'If a user account exists with the entered email, we will send an email with further instructions to reset the password.',
+          type: 'success',
+        });
+      },
+      onError: () => {
+        setNotification({
+          message: 'An error occurred. Please try again later',
+          type: 'error',
+        });
+      },
+    });
   };
 
   return (
@@ -64,8 +57,8 @@ const ForgotPasswordPopup: React.FC<ForgotPasswordProps> = ({ open, onClose }) =
         <FormProvider {...methods}>
           <Box component="form" onSubmit={handleSubmit(onSubmit)}>
             <TextFieldForm margin="normal" fullWidth name="email" label="Email" disabled={loading} />
-            {notification?.type === "success" && <Typography>{notification.message}</Typography>}
-            {notification?.type === "error" && <Typography>{notification.message}</Typography>}
+            {notification?.type === 'success' && <Typography>{notification.message}</Typography>}
+            {notification?.type === 'error' && <Typography>{notification.message}</Typography>}
           </Box>
         </FormProvider>
       </DialogContent>
