+++ src/stores/index.ts
@@ -1,5 +1,6 @@
-export { useLoaderStore } from "./loader.store";
-export { useUserStore } from "./user.store";
-export { usePageDataStore } from "./pageData.store";
-export { useNotificationStore } from "./notification.store";
-export { useErrorStore } from "./error.store";
+export { useLoaderStore } from './loader.store';
+export { useUserStore } from './user.store';
+export { usePageDataStore } from './pageData.store';
+export { useNotificationStore } from './notification.store';
+export { useErrorStore } from './error.store';
+export { useAuthStore } from './auth.store';
